test(pagination): add common test suite
