+++ tests/utils/normalizeSnapshot.ts
@@ -16,8 +16,8 @@ export function normalizeSnapshot(html: string) {
       // Vue renders extra whitespace between span elements
       .replace(/<\/span> <span/g, '</span><span')
       // Vue renders an empty comment for falsy v-if predicates
-      .replace(/(\s+)?<!---->/g, '') // Vue 2
-      .replace(/(\s+)?<!--v-if-->/g, '') // Vue 3
+      .replace(/(\s+)?<!---->(\s+)?/g, '') // Vue 2
+      .replace(/(\s+)?<!--v-if-->(\s+)?/g, '') // Vue 3
       // Vue renders extra whitespace after list elements
       .replace(/<\/ul> </g, '</ul><')
   );
